Camera focusing and stuff
Added a script that can be called from the editor that will perma-set the camera to a specific location.  Another script can be called to clear this setting. Also, the script call to start a tbs battle can be passed a boolean that tells it whether or not to clear the camera focus at the end of the battle. Not only will this make some cinematic stuff easier, but it can also be used to lock the screen to a particular area for smaller scale battles.

Also: added a toughness setting to enemies, and fixed a bug with battle scenes bailing out instantly.