Multi stage actions part 1
Grouping hits into hit groups. This will allow successive hits in one action. Not only will this allow later hits to be affected by stress and damage accumulated from previous hits, but will also allow for multi-stage animations where the enemy is shown to get hit (or not) more than once through the course of it.

This update applies the change to the current content, and also updates menus and initial loading to handle them. The next part will involve updating combat and animations.