Damage overhaul
Now stress and damage are both on a scale of 0 to 100. Everything else has been adjusted accordingly. The right status window in the battle scene is now at the top of the screen, and the status windows now show numbers rather than just text and symbols. The numbers all get higher as stress or damage is accumulated.

There was also an attempted first pass at rebalancing statistics for this new approach.