Damage/armor display numbers
To keep display numbers low, damage/armor values are multiplied by 10 before being used for combat math. So a damage value that reads 5 is actually a damage power of 50. This isn't the case for healing values. If a healing value is 100, it't displayed as double asterisks. Adjusted stress from healing to be more like stress from the same amount of damage.

Fixed a problem with the wrong targets being given for an action after manual moves. I thought I fixed this earlier but I actually didn't. Also made it so the previously chosen action isn't forgotten after a manual target.

Also re-enabled battler sound effects on taking damage.