Bounding movement in battle
Did a better job of calculating the bounding box for cursor movement when the camera is focused. Fixed a bug that previous attempts caused which allowed characters to walk outside of their movement field during manual movement. Also applied the bounding box to calculating move and action ranges in the first place, so nobody will move off the screen during camera focus.