Turn order changes
Reverting back to the old plan of turns rotating back and forth between forces after every actor acts. A round ends when all actors have acted. At the start of a battle or new round, force turn order is determined by skill rolls based on the Perception skill, with the highest roll among all members of a force applying to that force. Also, an actor with any existing damage when a battle starts will be inflicted with an appropriate amount of stress from the start.

Adjusted size of Stress meter in the battle actor status window, and made sure casting animations get mirrored on enemies.