Buckshot damage type
Added a new damage type, "buckshot." It divides the power of the damage type by 7 and then applies the result 7 times as 7 separate bypass/damage calculation. As such it can do huge damage to a limb but probably won't overpenetrate.

Fixed a bug with body part hit calculation that wasn't returning results after determining a limb hit. Fixed a bug with duplicate perception roll resolution resulting in an infinite loop.