Simple multihits and accuracy variance
added a multipleHit property to hits. This turns the hit into several hits during the combat math, but is displayed as a single row in the action info.

added an accuracyVariance property to hits, which acts as a random plus-or-minus adjustment on the accuracy bonus. it also adjusts the position of the hit animation accordingly.