Massive rangefinding overhauls (part 1)
Upon starting battle, tiles and events are collected into "tile runs," each of which is a collection of horizontally adjacent tiles with identical passability. When collision detecting for moving or attacking, these runs are passed into the collision algorithm as large rectangles to be checked against. This seems to massively improve the efficiency of these checks.