Massive rangefinding overhauls (part 3 final?)
Tile run construction at the start of battle now concatenates entire columns and rectangles of tiles into a single run if they all have the same passability. This is probably about as efficient as I can reasonably make that.

Also eliminated as many loops from the rangefinding process as possible. Everywhere there is a loop, it ticks up a number property of Game_Map by 1 for each run of the loop. This property is then used by the action processing queue to decide how much work is enough for a single frame. It seems to run pretty darn smoothly now.