Random stress effect
Randomized stress a bit. Half of its penalty is guaranteed but the other half is rolled. However, the maximum possible effect is more severe than before. The average effect is roughly the same as it was before. This should give even highly stressed characters something of a chance to come back.