Spending skill points
Main menu character status window has been separated into two windows. The lower one is now an interactable menu with all of the character's skills. They can be selected to spend skill points to upgrade them, or respec points to downgrade them and regain the skill points. Also added a sytem function to add skill points, which gives the points to the whole party, and also gives an equal amount of respec points.