Interface changes, and minor actions
Added a flag to actions that allows the current force to move another member after it is used. Added symbols to the character status skills list denoting which ability skills are the defense and reflex skills. Increased stress recovery. In battle targeting menu, pressing left or right jumps to the manual targeting option.