Some UI changes
A gray version of the body part damage indicator will appear if only a small amount of damage is taken. The round bonus meter is now labeled "Fo" for "Focus". Fixed a bug where a healing hit's action info display both types of healing even if one is not specified.