AI enhancements
An NPC targeting itself with its action which is either at high stress or has the "tactical" ai type will move to the furthest available cover, or as far as way from enemies as possible. When targeting something else, it will always move such that the target is just on the edge of the action's range.