Grouped actions in battle menus
Actions can be given an actionGroupName, and an actionGroupLevel. Actions with the same name will be "collapsed" into a single entry in the action menu in battle, using the name of the lowest level action, and (when not selected) the color and stress cost of the highest enabled action in the group. When selected, a horizontal menu will appear to the right of the actions menu, allowing the use of left/right to select actions of different levels.