Altered action groups in battle action menu
Now every action should have a group name and level. None of the levels need to be 1 for actions to get grouped. The action menu will show the name of whichever level action is currently selected, or the highest enabled one if the action isn't selected. No more than one action of a particular combination of name and level can be added, preventing duplicates.